Move Scorer to analysis tools
