+++ snorkel/slicing/utils.py
@@ -5,7 +5,7 @@ import torch
 from torch import nn
 
 from snorkel.classification.data import DictDataLoader
-from snorkel.classification.scorer import Scorer
+from snorkel.analysis import Scorer
 from snorkel.classification.snorkel_classifier import Operation, Task
 
 from .modules.slice_combiner import SliceCombinerModule
